+++ cardano-crypto-class/src/Cardano/Crypto/KES/Class.hs
@@ -77,6 +77,7 @@ import Cardano.Binary (Decoder, decodeBytes, Encoding, encodeBytes, Size, withWo
 import Cardano.Crypto.Util (Empty)
 import Cardano.Crypto.MLockedSeed
 import Cardano.Crypto.Hash.Class (HashAlgorithm, Hash, hashWith)
+import Cardano.Crypto.DSIGN.Class (failSizeCheck)
 
 class ( Typeable v
       , Show (VerKeyKES v)
@@ -319,33 +320,19 @@ encodeSignKeyKES = fmap encodeBytes . rawSerialiseSignKeyKES
 
 decodeVerKeyKES :: forall v s. KESAlgorithm v => Decoder s (VerKeyKES v)
 decodeVerKeyKES = do
-    bs <- decodeBytes
-    case rawDeserialiseVerKeyKES bs of
-      Just vk -> return vk
-      Nothing
-        | actual /= expected
-                    -> fail ("decodeVerKeyKES: wrong length, expected " ++
-                             show expected ++ " bytes but got " ++ show actual)
-        | otherwise -> fail "decodeVerKeyKES: cannot decode key"
-        where
-          expected = fromIntegral (sizeVerKeyKES (Proxy :: Proxy v))
-          actual   = BS.length bs
-{-# INLINEABLE decodeVerKeyKES #-}
+  bs <- decodeBytes
+  case rawDeserialiseVerKeyKES bs of
+    Just vk -> return vk
+    Nothing -> failSizeCheck "decodeVerKeyKES" "key" bs (sizeVerKeyKES (Proxy :: Proxy v))
+{-# INLINE decodeVerKeyKES #-}
 
 decodeSigKES :: forall v s. KESAlgorithm v => Decoder s (SigKES v)
 decodeSigKES = do
-    bs <- decodeBytes
-    case rawDeserialiseSigKES bs of
-      Just sig -> return sig
-      Nothing
-        | actual /= expected
-                    -> fail ("decodeSigKES: wrong length, expected " ++
-                             show expected ++ " bytes but got " ++ show actual)
-        | otherwise -> fail "decodeSigKES: cannot decode signature"
-        where
-          expected = fromIntegral (sizeSigKES (Proxy :: Proxy v))
-          actual   = BS.length bs
-{-# INLINEABLE decodeSigKES #-}
+  bs <- decodeBytes
+  case rawDeserialiseSigKES bs of
+    Just sig -> return sig
+    Nothing -> failSizeCheck "decodeSigKES" "signature" bs (sizeSigKES (Proxy :: Proxy v))
+{-# INLINE decodeSigKES #-}
 
 decodeSignKeyKES :: forall v s m. (UnsoundKESSignAlgorithm m v) => Decoder s (m (Maybe (SignKeyKES v)))
 decodeSignKeyKES = do