+++ cardano-crypto-class/CHANGELOG.md
@@ -1,3 +1,4 @@
+
 ## 2.1.0.1
 
 * Remove `development` flag: #372