+++ cardano-crypto-class/CHANGELOG.md
@@ -1,3 +1,5 @@
+# Changelog for `cardano-crypto-class`
+
 ## 2.1.1.0
 * Included bindings to `blst` library to enable operations over curve BLS12-381